+++ internal/state/mailbox_search.go
@@ -59,7 +59,7 @@ func (m *Mailbox) Search(ctx context.Context, keys []command.SearchKey, decoder 
 	if err := parallel.DoContext(ctx, parallelism, msgCount, func(ctx context.Context, i int) error {
 		defer async.HandlePanic(m.state.panicHandler)
 
-		msg, ok := m.snap.messages.getWithSeqID(imap.SeqID(i + 1))
+		msg, ok := m.snap.messages.getWithSeqID(imap.SeqID(uint32(i + 1)))
 		if !ok {
 			return nil
 		}
